requirements.txt
fastapi==0.68.0
pymongo==3.12.0
pytest==6.2.4
python-multipart==0.0.5
uvicorn==0.13.4
cffi==1.14.6
cryptography==3.4.7
ecdsa==0.17.0
pyasn1==0.4.8
pycparser==2.20
python-jose==3.3.0
rsa==4.7.2
bcrypt==3.2.0
passlib==1.7.4
pytest-cov==2.12.1